Began production on an NPC generator class that incorporates instances of race, class and character so that a single NPC can be instantiated with one or two lines of code, rather than individually loading in every parameter.
Created and passed several tests on a method that randomly produces an NPC's race. Two separate methods for the race's file name (half-orc.json) and the race's string name (half-orc) were created so that loading with file name can be done for pulling data, and so printing information about an NPC to the player can be done without the ".json" removing from immersion.