Added a data folder for monsters.
Created a new class of Monster. Created new methods for stat generation within monsters that accept longs as a parameter, as opposed to ints within Character. This is because the data that is within the .json files is stored as a long and will throw ClassCastExceptions whenever the int method is attempted to be used within it.

Created and passed tests surrounding the idea that hit points are determined by the monster's average HP (goblins being 7), and their constitution affecting this score. This can create variation within monsters, making some stronger and some weaker. Tests were also created and passed surrounding the fact that each monster's stats are individually rolled and affected by their average score -- for instance, a goblin's average dexterity is 14, making their modifier +2 that is added when their stats are rolled.

This will create much diversity within the monsters that are found in the game, some being much more difficult, or at least vastly different, than the stereotypical type found within the monster manual.