Heavy tweaking on the structure of NPCGenerator. Massive overhaul on the generateNPC() method, which needed explicit setting for stats such as armor class and hit points for them to not return a value of 0. Extensive testing has confirmed that using the generateNPC() method will set the armor class and hit points to what they should be -- depending on constitution and hit die and being within their proper ranges.
Tweaked the structure of the racialAbility methods so that humans do not cause a test to fail, even though they lack racial abilities. Asking if a human has a racial ability, such as Darkvision, will still return false.

Moved some methods from NPCGenerator to the Monster class. I will eventually add a method that adds monster races, such as goblins, to be non-hostile NPCs, but that is further down the line.